+++ src/components/register/ClassName.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const ClassName = () => {
+  return (
+    <div>ClassName</div>
+  )
+}
+
+export default ClassName
